+++ src/main/java/forestry/core/data/ForestryBiomeTagsProvider.java
@@ -12,12 +12,15 @@ import thedarkcolour.modkit.data.MKTagsProvider;
 public class ForestryBiomeTagsProvider {
 	public static void addTags(MKTagsProvider<Biome> tags, HolderLookup.Provider lookup) {
 		// Climates
+		tags.tag(ForestryTags.Biomes.ICY_TEMPERATURE)
+				.add(Biomes.FROZEN_OCEAN)
+				.add(Biomes.DEEP_FROZEN_OCEAN)
+		;
 		tags.tag(ForestryTags.Biomes.COLD_TEMPERATURE)
 				.addTag(BiomeTags.IS_END)
 				.add(Biomes.COLD_OCEAN)
 				.add(Biomes.DEEP_OCEAN)
 				.add(Biomes.DEEP_COLD_OCEAN)
-				.add(Biomes.DEEP_FROZEN_OCEAN)
 		;
 		tags.tag(ForestryTags.Biomes.WARM_TEMPERATURE)
 				.add(Biomes.WOODED_BADLANDS)